requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# EEG Processing
mne>=1.6.0
braindecode>=0.8.0
pyedflib>=0.1.30

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Signal Processing
pywavelets>=1.4.0

# Config Management
pyyaml>=6.0
omegaconf>=2.3.0

# Logging & Progress
tqdm>=4.66.0
tensorboard>=2.15.0
rich>=13.0.0

# System Info
psutil>=5.9.0
pynvml>=11.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Stats
statsmodels>=0.14.0

# Utilities
requests>=2.31.0
einops>=0.7.0
timm>=0.9.0
